requirements.txt
flask
flask-sqlalchemy
flask-restful
sqlalchemy-migrate
flask-cors
names